core: add a big_space grid and compatible camera
